pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "BLDFM"
version = "1.0.0-dev"
description = "Boundary Layer Footprint Dispersion Model"
authors = [
    { name = "Mark Schlutow", email = "mschlutow@bgc-jena.mpg.de" },
    { name = "Ray Chew", email = "chew@caltech.edu" },
    { name = "Mathias Göckede", email = "mgoeck@bgc-jena.mpg.de" },

]
license = { text = "MIT" }
dependencies = [
    "numpy",
    "matplotlib",
    "scipy",
    "numba",
    "pyfftw",
    "pyyaml",
    "xarray",
    "netcdf4",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "sphinx",
    "black",
]
plotting = [
    "contextily",
    "windrose",
    "plotly",
]

[project.scripts]
bldfm = "bldfm.cli:main"

[tool.pytest.ini_options]
markers = [
    "unit: Fast tests that do not run the solver",
    "integration: Tests that run the solver (single, timeseries, multitower, I/O)",
    "parallel: Tests that use ProcessPoolExecutor",
    "plotting: Tests for matplotlib/plotly plotting functions",
]
